Feat: Endpoint para Historial de Predicciones y Refactor de Paginación

+++ src/main/java/com/equipo_38/flight_on_time/dto/ResponsePageDTO.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public record ResponsePageDTO<T>(
         List<T> content,
-        int size
+        Long size
 ) {
 }
